new try

+++ exercises.c
@@ -139,13 +139,13 @@ int parentesisBalanceados(char *cadena)
       char c = cadena[i];  // variable para facilitar comparaciones y operaciones
       if(c == '(' || c == '{' || c == '[') // si es apertura
       {
-         push(aux, (void*)(long) c );   // se inserta en la pila. char --> long --> void* | caracter --> valor ascii --> puntero a void, funcion propuesta requiere void*;
+         push(aux, (void*) c );   // se inserta en la pila. char --> long --> void* | caracter --> valor ascii --> puntero a void, funcion propuesta requiere void*;
          // char c = numero pequeno , void* -> direccion de memoria , (void*)c --> (void*) 21 --> invalido , "21" como ejemplo.
       }   
       else if(c == ')' || c == '}' || c == ']')   //  si es cierre
       {
          if(top(aux) == NULL) return 0; // si no hay apertura, retorna falso.
-         char topc = (char) (long) top(aux); // void* --> long --> char . variable de comparacion
+         char topc = (char) top(aux); // void* --> long --> char . variable de comparacion
          pop(aux);  // ya se guarda el top en topc, por lo que se elimina para comparar el siguiente en la proxima iteracion.
          if( (c == ')' && topc != '(') || (c == '}' && topc != '{') || (c == ']' && topc != '[') ) return 0; // si apertura no coincide con cierre retorna falso.
       }
